#99 minor: moved methods for clarity, expecting to do more tidying later
